Megaseed Vendor's Dionapod + Cotton, Re-Organized and, Re-Alphabetized (#11989)
* Megaseed Vendor Alphabetical Sorting

Fixes the placement of the Dionapod in the vendor. Unlike every other seed these weren't in alphabetical order, so simply helped them get where they belong :)

* MegaSeed Vendor Cotton Reorder

## About The Pull Request

ANOTHER ONE IN THE WRONG ORDER!! Similar to the Dianapods which were *not* in alphabetical order, while _most_ other things were, this is an attempt to right a wrong and put <ins>Cotton</ins> back where it belongs alphabetically in the MegaSeed Vendor. It'll place it, rather then between *A*pple, and *B*anana, it in its cozy new home, right between *C*orn and *D*ionapod. hopefully fully organizing the vendor~

## Why Its Good For The Game

Well, _especially_ if DIanapods are going to be moved, then it'd be mighty silly to just have *one* plant be out of order. But for the same reason I wanted them to be reorganized, the Megaseed Vendor deserves to be neat, and orderly. This helps that come true.

## Changelog

:cl:
Fix: MegaSeed Vendor's Cotton gets Alphabetically Organized
:/cl: