Commit 4 (explosive spear warcries, ventriloquist and custom vendors)
TGUIfied and added filters for explosive spear warcries, ventriloquist and custom vendors

Added description to ventriloquist dummy explaining how to use it since it wasn't obvious.

On the custom vendors, tgufied it so it's a list and you select what you want to edit instead of it being one at a time.