Full functionality
Moves the given code to only function in pockets. Stops simply grabbing syringes from hands from stabbing you. Now also checks for if the user pick pocketing is a carbon to prevent any weird cases. Further checks for enhanced retrieval gloves to at least give sec an option to not get jabbed. Might be a good idea to make it a type of glove that works instead of just enhanced retrievals.