Sharps Hazard - Syringes embed when stripped from pockets (#12991)
* Update strippable.dm

Makes basic syringe embed mechanics to searching pockets filled with syringe.

* Full functionality

Moves the given code to only function in pockets. Stops simply grabbing syringes from hands from stabbing you. Now also checks for if the user pick pocketing is a carbon to prevent any weird cases. Further checks for enhanced retrieval gloves to at least give sec an option to not get jabbed. Might be a good idea to make it a type of glove that works instead of just enhanced retrievals.

* PR Review Compliance

Now compliant with all the PR reviews. Changes the types of gloves that are resistant to the syringes, and makes a few of such. Also adds functionality to fullstripping jumpsuits. Piercing syringes also go through the specific gloves when fullstripping.

* Bacon Compliance

Adds new proc that gets called in start_unequip. Used for checking items for effects when being stripped. Moved relevant effects of embedding to the syringe specific version of this proc. Uniforms now also run this check for items in the pockets when being stripped.

* Formatting Fixes

Makes things cleaner and look nicer with better functionality.