Overhauls blocking + a couple of new toys to block with.  (#13123)
* Variable renames

* shields with NO_DROP call shatter() if stam is too high

* SORDs can't block

Well they sort of can, but it's a really bad idea to try.

* Beepsky smash no longer gives NO_BLOCK

* removes blockbreak() from humans

redundant proc better handled with a COOLDOWN macro

* Supermatter sword comment

* Defines blocking cooldown under combat defines

* Updates simplemob variable usage

* Core blocking changes

* fix block angles

* Debugging

I can't prevent a block from within on_block, has to be in hit_reaction

* Three new flags

* Energy weapons

block projectiles + unblockable flags

* fixes SORD, gives it active blocking

* Sticks and Scythes

Wizard and Chaplain staves, Chaplain Scythe, Bo staves

* shields

* replica katana

Is now a weak katana, but a usable weapon

* Misc.

* shield tweak and warning

* removes unused variable "final_block_chance"

* Carbon Fiber and other sabre shenanigans

* spears, knives and kinetic crusher

* fixes as I gather testing evidence

* undo basemap

* I forgot the mop

* Teleprod

* no stamcrit when blocking

* shield consistency

* FALSE not 0

* Missing flags, missing wield checks, revert nightmare

* Removes pitchfork

* Revert "Removes pitchfork"

This reverts commit 6f74e8916a5b1a14ad4d80532ad0019e00d4db4f.

* Fixes pitchfork

Nevermind the linter throws a shitfit if I try to remove it and I'm not fucking with that.

* Polish, fixes and flavor messages

* whoops

* contain the dangerous flags

* fix stun baton

* Tweaks after first test

* cheaper energy shields

* Fixes shield icons

* updates ashwalker camp

* Adds backsprites for shields that now fit on backs

* Fix shields by using AND not OR

* Update items.dm

* Fix chaplain bo staff

* block_power tweaks after testing, fixes combat info

* Touches up chaplain holy staff, you can now turn the barrier on and off.

* Fixes energy shields

Non functional recharging capabilities have been fixed and the sound that plays when the shield is charged is now audible.

* Hanzo steel brought down to match the other station katanas

* Improve consistency of maximum blocking stamina damage

The absolute maximum under the stars aligning of stamina damage someone can take is now 69, if they had 34 when they got hit and took the maximum stamina output of 35