Reduces vampire feed cooldown and adds grab check.
Reduces the cooldown time for the vampire feed power to make it less punishing.

Adds a check to see if the vampire successfully grabs the target before applying immobilization traits to both parties.
This is so the vampire power does not activate and apply traits without the grab happening.
If the grab fails, it alerts the vampire, deactivates the power, and returns false.