Redeems security voucher into a gear box.
Changes the security voucher redemption process to spawn a box containing the selected equipment.

This ensures that the redeemed items are contained and easily managed, preventing them from being scattered around the vendor location. The spawned box is then given to the redeemer. If they can't hold it, it is dropped on the floor at their location.