Implements nuke code request auto-approval.
Adds an auto-approval mechanism for nuke requests with admin veto.

After a request is made, a timer starts. If no admin rejects the request within a set time, the code is automatically generated and applied to all nukes.

Admins can reject requests, cancelling the auto-approval and notifying the requester.
A station announcement is made after auto-approval.

Adds station announcement type.