Adds non-oxygenated fire creation and adds gas reaction to fhf
Implements hydrogen fuel as a new flammable gas.

This change introduces hydrogen fuel as a new gas type, defining its properties and behavior within the atmospherics system.
It can be used as a propellant for orbital thrusters, and is mildly toxic when inhaled.
The combustion of hydrogen fuel produces water vapor.