Adds tutorial notes to orbital controls
Adds sticky notes to the orbital height control and fuel supply computer to provide players with more information on how to operate them.

Adds some extra flavor text to the nitrium fuel description.

Prevents thrust particles from displaying if the thruster backend does not exist, or it is out of fuel.